Test heroku

+++ src/main/java/com/kaidash/mobileapplicationsservice/service/ApplicationServiceImpl.java
@@ -31,9 +31,9 @@ public class ApplicationServiceImpl implements ApplicationService {
     public Application findApplicationById(int applicationId) {
 
         Optional<Application> applicationOptional = applicationRepository.findById(applicationId);
-//        if(applicationOptional.isEmpty()) {
-//            throw new EntityNotFoundException(String.format("Application id [%s] not found.", applicationId));
-//        }
+        if(applicationOptional.isEmpty()) {
+            throw new EntityNotFoundException(String.format("Application id [%s] not found.", applicationId));
+        }
         return applicationOptional.get();
     }
 
@@ -60,9 +60,9 @@ public class ApplicationServiceImpl implements ApplicationService {
     public ApplicationProcessingResponse deleteApplicationById(int applicationId) {
 
         Optional<Application> applicationOptional = applicationRepository.findById(applicationId);
-        //if(applicationOptional.isEmpty()) {
-        //    throw new EntityNotFoundException(String.format("Application id [%s] not found.", applicationId));
-        //}
+        if(applicationOptional.isEmpty()) {
+            throw new EntityNotFoundException(String.format("Application id [%s] not found.", applicationId));
+        }
 
         applicationRepository.deleteById(applicationId);
